Consolidate relative time formatting
Centralize relative time/date formatting into a shared utility and reuse it across the app and example plugins. Also add focus-visible styles for reset-styled branch/PR buttons.
