+++ app/styles/app.css
@@ -2224,6 +2224,11 @@ body,
   background: var(--bg-hover);
 }
 
+.pr-link-badge:focus-visible {
+  outline: 2px solid var(--accent);
+  outline-offset: 2px;
+}
+
 .pr-link-badge .pr-draft-indicator {
   font-size: 9px;
   padding: 1px 4px;
@@ -4825,6 +4830,12 @@ body,
   all: unset;
   cursor: pointer;
   display: inline-flex;
+}
+
+.pr-review-branch .branch-link:focus-visible {
+  outline: 2px solid var(--accent);
+  outline-offset: 2px;
+  border-radius: 6px;
 }
 
 .pr-review-branch .branch-link code {
